r5619/sofa-dev : FIX compilation on windows (try...) Remove some warnings (try...)
 //partial sync (4/5 changes)//

+++ modules/sofa/component/collision/LineLocalMinDistanceFilter.cpp
@@ -49,7 +49,7 @@ void LineInfo::buildFilter(unsigned int edge_index)
 
     bool debug=false;
 
-    if (edge_index==-1)
+    if ((int)edge_index==-1)
         debug=true;
 
 
@@ -157,7 +157,7 @@ bool LineInfo::validate(const unsigned int edge_index, const defaulttype::Vector
 {
     bool debug=false;
 
-    if (edge_index==-1)
+    if ((int)edge_index==-1)
         debug=true;
